+++ sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/TPCDSQueryBenchmark.scala
@@ -17,19 +17,7 @@
 
 package org.apache.spark.sql.execution.benchmark
 
-import scala.util.Try
-
-import org.apache.spark.SparkConf
-import org.apache.spark.benchmark.Benchmark
-import org.apache.spark.internal.Logging
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.TPCDSSchema
-import org.apache.spark.sql.catalyst.catalog.HiveTableRelation
-import org.apache.spark.sql.catalyst.plans.logical.SubqueryAlias
-import org.apache.spark.sql.catalyst.util._
-import org.apache.spark.sql.catalyst.util.DateTimeConstants.NANOS_PER_SECOND
-import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types.StructType
 
 /**
@@ -45,92 +33,17 @@ import org.apache.spark.sql.types.StructType
  *      Results will be written to "benchmarks/TPCDSQueryBenchmark-results.txt".
  * }}}
  */
-object TPCDSQueryBenchmark extends SqlBasedBenchmark with Logging {
+object TPCDSQueryBenchmark extends TPCBenchmarkUtils {
 
-  override def getSparkSession: SparkSession = {
-    val conf = new SparkConf()
-      .setMaster(System.getProperty("spark.sql.test.master", "local[1]"))
-      .setAppName("test-sql-context")
-      .set("spark.sql.parquet.compression.codec", "snappy")
-      .set("spark.sql.shuffle.partitions", System.getProperty("spark.sql.shuffle.partitions", "4"))
-      .set("spark.driver.memory", "3g")
-      .set("spark.executor.memory", "3g")
-      .set("spark.sql.autoBroadcastJoinThreshold", (20 * 1024 * 1024).toString)
-      .set("spark.sql.crossJoin.enabled", "true")
-      .set("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-      .set("spark.kryo.registrationRequired", "true")
-
-    SparkSession.builder.config(conf).getOrCreate()
-  }
-
-  val tables = Seq("catalog_page", "catalog_returns", "customer", "customer_address",
-    "customer_demographics", "date_dim", "household_demographics", "inventory", "item",
-    "promotion", "store", "store_returns", "catalog_sales", "web_sales", "store_sales",
-    "web_returns", "web_site", "reason", "call_center", "warehouse", "ship_mode", "income_band",
-    "time_dim", "web_page")
-
-  def setupTables(dataLocation: String, tableColumns: Map[String, StructType]): Map[String, Long] =
-    tables.map { tableName =>
-      spark.sql(s"DROP TABLE IF EXISTS $tableName")
-      val options = Map("path" -> s"$dataLocation/$tableName")
-      spark.catalog.createTable(tableName, "parquet", tableColumns(tableName), options)
-      // Recover partitions but don't fail if a table is not partitioned.
-      Try {
-        spark.sql(s"ALTER TABLE $tableName RECOVER PARTITIONS")
-      }.getOrElse {
-        logInfo(s"Recovering partitions of table $tableName failed")
-      }
-      tableName -> spark.table(tableName).count()
-    }.toMap
-
-  def runTpcdsQueries(
-      queryLocation: String,
-      queries: Seq[String],
-      tableSizes: Map[String, Long],
-      nameSuffix: String = ""): Unit = {
-    queries.foreach { name =>
-      val queryString = resourceToString(s"$queryLocation/$name.sql",
-        classLoader = Thread.currentThread().getContextClassLoader)
-
-      // This is an indirect hack to estimate the size of each query's input by traversing the
-      // logical plan and adding up the sizes of all tables that appear in the plan.
-      val queryRelations = scala.collection.mutable.HashSet[String]()
-      spark.sparkContext.setJobGroup(name, s"$name:\n$queryString", true)
-      spark.sql(queryString).queryExecution.analyzed.foreach {
-        case SubqueryAlias(alias, _: LogicalRelation) =>
-          queryRelations.add(alias.name)
-        case LogicalRelation(_, _, Some(catalogTable), _) =>
-          queryRelations.add(catalogTable.identifier.table)
-        case HiveTableRelation(tableMeta, _, _, _, _) =>
-          queryRelations.add(tableMeta.identifier.table)
-        case _ =>
-      }
-      val numRows = queryRelations.map(tableSizes.getOrElse(_, 0L)).sum
-      val benchmark = new Benchmark(s"TPCDS Snappy", numRows, 2, output = output)
-      benchmark.addCase(s"$name$nameSuffix") { _ =>
-        spark.sql(queryString).noop()
-      }
-      benchmark.run()
-    }
-  }
-
-  private def filterQueries(
-      origQueries: Seq[String],
-      queryFilter: Set[String],
-      nameSuffix: String = ""): Seq[String] = {
-    if (queryFilter.nonEmpty) {
-      if (nameSuffix.nonEmpty) {
-        origQueries.filter { name => queryFilter.contains(s"$name$nameSuffix") }
-      } else {
-        origQueries.filter(queryFilter.contains)
-      }
-    } else {
-      origQueries
-    }
-  }
+  override val tables: Seq[String] = Seq("catalog_page", "catalog_returns", "customer",
+    "customer_address", "customer_demographics", "date_dim", "household_demographics",
+    "inventory", "item", "promotion", "store", "store_returns", "catalog_sales",
+    "web_sales", "store_sales", "web_returns", "web_site", "reason", "call_center",
+    "warehouse", "ship_mode", "income_band", "time_dim", "web_page")
+  override val queryType: String = "TPCDS"
 
   override def runBenchmarkSuite(mainArgs: Array[String]): Unit = {
-    val benchmarkArgs = new TPCDSQueryBenchmarkArguments(mainArgs)
+    val benchmarkArgs = new TPCQueryBenchmarkArguments(mainArgs)
 
     // List of all TPC-DS v1.4 queries
     val tpcdsQueries = Seq(
@@ -165,25 +78,10 @@ object TPCDSQueryBenchmark extends SqlBasedBenchmark with Logging {
 
     val tableSizes = setupTables(benchmarkArgs.dataLocation,
       TPCDSSchemaHelper.getTableColumns)
-    if (benchmarkArgs.cboEnabled) {
-      spark.sql(s"SET ${SQLConf.CBO_ENABLED.key}=true")
-      spark.sql(s"SET ${SQLConf.PLAN_STATS_ENABLED.key}=true")
-      spark.sql(s"SET ${SQLConf.JOIN_REORDER_ENABLED.key}=true")
-      spark.sql(s"SET ${SQLConf.HISTOGRAM_ENABLED.key}=true")
+    configureCbo(benchmarkArgs)
 
-      // Analyze all the tables before running TPCDS queries
-      val startTime = System.nanoTime()
-      tables.foreach { tableName =>
-        spark.sql(s"ANALYZE TABLE $tableName COMPUTE STATISTICS FOR ALL COLUMNS")
-      }
-      logInfo("The elapsed time to analyze all the tables is " +
-        s"${(System.nanoTime() - startTime) / NANOS_PER_SECOND.toDouble} seconds")
-    } else {
-      spark.sql(s"SET ${SQLConf.CBO_ENABLED.key}=false")
-    }
-
-    runTpcdsQueries(queryLocation = "tpcds", queries = queriesV1_4ToRun, tableSizes)
-    runTpcdsQueries(queryLocation = "tpcds-v2.7.0", queries = queriesV2_7ToRun, tableSizes,
+    runTpcQueries(queryLocation = "tpcds", queries = queriesV1_4ToRun, tableSizes)
+    runTpcQueries(queryLocation = "tpcds-v2.7.0", queries = queriesV2_7ToRun, tableSizes,
       nameSuffix = nameSuffixForQueriesV2_7)
   }
 }
